pyproject.toml
[tool.poetry]
name = "telegram-feedback-bot"
version = "0.1.0"
description = ""
authors = ["Pavel Pavlovskiy <xpavlus@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "telegram_feedback_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"
aiogram = "3.0.0b7"
fluent-runtime = "0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
